refactor: 상품 등록, 수정 리팩토링
